requirements.dev.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.dev.txt requirements.dev.in
#
black==21.12b0
    # via -r requirements.dev.in
click==8.1.7
    # via black
colorama==0.4.6
    # via
    #   click
    #   pytest
coverage==7.3.2
    # via -r requirements.dev.in
flake8==3.9.2
    # via -r requirements.dev.in
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via -r requirements.dev.in
mccabe==0.6.1
    # via flake8
mypy-extensions==1.0.0
    # via black
packaging==23.2
    # via pytest
pathspec==0.11.2
    # via black
platformdirs==4.1.0
    # via black
pluggy==1.3.0
    # via pytest
pycodestyle==2.7.0
    # via flake8
pyflakes==2.3.1
    # via flake8
pytest==7.4.3
    # via
    #   -r requirements.dev.in
    #   pytest-django
pytest-django==4.7.0
    # via -r requirements.dev.in
tomli==1.2.3
    # via black
typing-extensions==4.8.0
    # via black
